new update
java 18
Loading screen DEMO
